+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+[1.0.3]
+
+Added hideSuggestionsOnInit optional parameter to hide or show all suggestions when widget is created. 
+Default value is true. Also tileMinHeight and tileMaxHeight parameters added.
+
 [1.0.2]
 
 Fixed some issues about providing callback function as a parameter.